+++ test/com/cloudogu/ces/cesbuildlib/K3dTest.groovy
@@ -34,9 +34,9 @@ class K3dTest extends GroovyTestCase {
         sut.deleteK3d()
 
         // then
-        assertThat(scriptMock.actualShStringArgs[6].trim()).contains("k3d registry delete citest-")
-        assertThat(scriptMock.actualShStringArgs[7].trim()).contains("k3d cluster delete citest-")
-        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(8)
+        assertThat(scriptMock.actualShStringArgs[9].trim()).contains("k3d registry delete citest-")
+        assertThat(scriptMock.actualShStringArgs[10].trim()).contains("k3d cluster delete citest-")
+        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(11)
     }
 
     void testKubectl() {
@@ -66,7 +66,33 @@ class K3dTest extends GroovyTestCase {
         assertThat(scriptMock.actualShStringArgs[3].trim()).matches("k3d registry create citest-[0-9a-f]+ --port 54321")
         assertThat(scriptMock.actualShStringArgs[4].trim()).startsWith("k3d cluster create citest-")
         assertThat(scriptMock.actualShStringArgs[5].trim()).startsWith("k3d kubeconfig merge citest-")
-        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(6)
+        assertThat(scriptMock.actualShStringArgs[6].trim()).startsWith("echo \"Using credentials: cesmarvin-setup\"")
+        assertThat(scriptMock.actualShStringArgs[7].trim()).startsWith("sudo KUBECONFIG=leWorkspace/.k3d/.kube/config kubectl delete secret dogu-cloudogu-com || true")
+        assertThat(scriptMock.actualShStringArgs[8].trim()).startsWith("sudo KUBECONFIG=leWorkspace/.k3d/.kube/config kubectl create secret generic dogu-cloudogu-com --from-literal=username=\"null\" --from-literal=password=\"null\"")
+        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(9)
+    }
+
+    void testStartK3dWithCustomCredentials() {
+        def workspaceDir = "leWorkspace"
+        def k3dVer = "4.4.7"
+
+        def scriptMock = new ScriptMock()
+        scriptMock.expectedShRetValueForScript.put('echo -n $(python3 -c \'import socket; s=socket.socket(); s.bind(("", 0)); print(s.getsockname()[1]); s.close()\');'.toString(), "54321")
+
+        K3d sut = new K3d(scriptMock, "${workspaceDir}", "path", "myBackendCredentialsID")
+
+        sut.startK3d()
+
+        assertThat(scriptMock.actualShStringArgs[0].trim()).isEqualTo("rm -rf ${workspaceDir}/.k3d".toString())
+        assertThat(scriptMock.actualShStringArgs[1].trim()).isEqualTo("mkdir -p ${workspaceDir}/.k3d/bin".toString())
+        assertThat(scriptMock.actualShStringArgs[2].trim()).isEqualTo("curl -s https://raw.githubusercontent.com/rancher/k3d/main/install.sh | TAG=v${k3dVer} K3D_INSTALL_DIR=${workspaceDir}/.k3d/bin bash -s -- --no-sudo".toString())
+        assertThat(scriptMock.actualShStringArgs[3].trim()).matches("k3d registry create citest-[0-9a-f]+ --port 54321")
+        assertThat(scriptMock.actualShStringArgs[4].trim()).startsWith("k3d cluster create citest-")
+        assertThat(scriptMock.actualShStringArgs[5].trim()).startsWith("k3d kubeconfig merge citest-")
+        assertThat(scriptMock.actualShStringArgs[6].trim()).startsWith("echo \"Using credentials: myBackendCredentialsID\"")
+        assertThat(scriptMock.actualShStringArgs[7].trim()).startsWith("sudo KUBECONFIG=leWorkspace/.k3d/.kube/config kubectl delete secret dogu-cloudogu-com || true")
+        assertThat(scriptMock.actualShStringArgs[8].trim()).startsWith("sudo KUBECONFIG=leWorkspace/.k3d/.kube/config kubectl create secret generic dogu-cloudogu-com --from-literal=username=\"null\" --from-literal=password=\"null\"")
+        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(9)
     }
 
     void testBuildAndPush() {
@@ -101,7 +127,7 @@ class K3dTest extends GroovyTestCase {
         sut.buildAndPushToLocalRegistry(imageName, imageTag)
 
         // then
-        assertThat(scriptMock.actualShStringArgs[6].trim()).isEqualTo("image pushed".toString())
-        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(7)
+        assertThat(scriptMock.actualShStringArgs[9].trim()).isEqualTo("image pushed".toString())
+        assertThat(scriptMock.actualShStringArgs.size()).isEqualTo(10)
     }
 }